+++ R/timeProp.R
@@ -29,8 +29,10 @@
 #' @param avg.time This defines the time period to average to. Can be `"sec"`,
 #'   `"min"`, `"hour"`, `"day"`, `"DSTday"`, `"week"`, `"month"`, `"quarter"` or
 #'   `"year"`. For much increased flexibility a number can precede these options
-#'   followed by a space. For example, a timeAverage of 2 months would be
-#'   `period = "2 month"`.
+#'   followed by a space. For example, an average of 2 months would be
+#'   `avg.time = "2 month"`. In addition, `avg.time` can equal `"season"`, in
+#'   which case 3-month seasonal values are calculated with spring defined as
+#'   March, April, May and so on.
 #'
 #'   Note that `avg.time` when used in `timeProp` should be greater than the
 #'   time gap in the original data. For example, `avg.time = "day"` for hourly
@@ -108,7 +110,7 @@ timeProp <- function(
   mydata <- checkPrep(mydata, vars, c(type, proportion), remove.calm = FALSE)
 
   # cut data
-  mydata <- cutData(mydata, c(type, proportion))
+  mydata <- cutData(mydata, c(type, proportion), ...)
 
   # time zone of input data
   tzone <- attr(mydata$date, "tzone")
@@ -117,15 +119,33 @@ timeProp <- function(
   group_1 <- c("xleft", "xright", type)
   group_2 <- c(type, "xleft", "xright", proportion)
 
+  # calculate left and right extremes of each bar, add the most common
+  # non-zero time interval to left to get right
+  if (avg.time == "season") {
+    if (any(c("season", "seasonyear") %in% type)) {
+      cli::cli_abort(
+        "In {.fun openair::timeProp}, {.arg type} and {.arg avg.time} cannot both be 'season'."
+      )
+    }
+    results <- mydata |>
+      cutData(type = "seasonyear") |>
+      dplyr::mutate(xleft = min(.data$date), .by = c("seasonyear", type)) |>
+      dplyr::mutate(
+        xright = .data$xleft + median(diff(.data$xleft)[diff(.data$xleft) != 0])
+      ) |>
+      dplyr::select(-dplyr::any_of("seasonyear"))
+  } else {
+    results <-
+      dplyr::mutate(
+        mydata,
+        xleft = as.POSIXct(cut(.data$date, avg.time), tz = tzone),
+        xright = .data$xleft + median(diff(.data$xleft)[diff(.data$xleft) != 0])
+      )
+  }
+
   # summarise by proportion, type etc
   results <-
-    mydata |>
-    # calculate left and right extremes of each bar, add the most common
-    # non-zero time interval to left to get right
-    dplyr::mutate(
-      xleft = as.POSIXct(cut(.data$date, avg.time), tz = tzone),
-      xright = .data$xleft + median(diff(.data$xleft)[diff(.data$xleft) != 0])
-    ) |>
+    results |>
     # calculate group mean
     dplyr::mutate(
       mean_value = mean(.data[[pollutant]], na.rm = TRUE),