+++ NEWS.md
@@ -25,6 +25,10 @@
     - `sub` can now be defined via `...`; set `sub = NA` to remove the text annotation which appears by default at the bottom of a `timeProp()` plot.
 
     - Gained the `key` argument to remove a legend.
+    
+    - `"season"` is now a permitted `avg.time` option in `timeProp()`, better aligning it with the options in `timeAverage()`.
+    
+    - `...` is now correctly passed to `cutData()` when using `type`/`proportion`.
 
 - `trajPlot()` and `trajLevel()` have gained the `grid.nx` and `grid.ny` arguments which can be used to control the number of ticks on the coordinate grid, or remove it altogether.